fix(tests): single shared bootstrap; idempotent SerializationThreadWorker.Exit
The UOContent.Tests host hung ~2.5 min at shutdown and crashed when the full
suite ran. Both collection fixtures (UOContentFixture, PathfindingTestFixture)
each ran the full process-global ModernUO bootstrap. World.Load() is guarded to
run once per process, so the second fixture's World.Load() was a no-op and did
NOT respawn the serialization workers -- but World.ExitSerializationThreads() is
not guarded, so the second fixture called Exit() on workers whose threads had
already terminated, blocking forever on _stopEvent.WaitOne() in the fixture
constructor. The first 258 tests ran, then the host stalled and was killed.

(a) Make SerializationThreadWorker.Exit() idempotent so a second Exit() is a
    safe no-op instead of a permanent block. Only the owning (main) thread calls
    Exit(), so no synchronization is needed; the single-call production path is
    unchanged.

(b) Centralize the test bootstrap: TestServerBootstrap.EnsureInitialized() runs
    the superset global init exactly once per process (lock + once-flag). Both
    collection fixtures delegate to it and no longer tear down global state. Add
    [assembly: CollectionBehavior(DisableTestParallelization=true)] so
    collections never overlap.

Full suite now: 418/418 pass in ~3s with a clean exit (was 258 partial then a
2.5-min hang and host crash).
